+++ packages/core/js-client/src/fetchers/node.ts
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2023 Fluence Labs Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { createRequire } from 'module';
+import { Buffer } from 'buffer';
+import fs from 'fs';
+
+export async function fetchResource(packageName: string, assetPath: string, version: string) {
+    const require = createRequire(import.meta.url);
+
+    const file = await new Promise<Buffer>((resolve, reject) => {
+        // Cannot use 'fs/promises' with current vite config. This module is not polyfilled by default.
+        const workerFilePath = require.resolve(packageName + assetPath);
+        fs.readFile(workerFilePath, (err, data) => {
+            if (err) {
+                reject(err);
+                return;
+            }
+            resolve(data);
+        });
+    });
+    return new Response(file, {
+        headers: {
+            'Content-type':
+                assetPath.endsWith('.wasm')
+                    ? 'application/wasm'
+                    : assetPath.endsWith('.js')
+                        ? 'application/javascript'
+                        : 'application/text'
+        }
+    });
+}
